Move TaskSiloLimit to taskworker
